Move react-scripts as a devDependency
This is a small change to *help* Snyk to work properly. Apparently, react-scripts has a large number of dependencies and is causing a large dependency tree. By moving to the devDependencies, the tree generated by Snyk is pruned because it does not scan this section.

It, anyway, makes sense to move react-scripts as a devDepencency, because it's not used for coding (i.e. not imported in any file under src/), but it is just part of the dev/build framework.